requirements.txt
streamlit==1.16.0
google-generativeai==0.2.1
python-dotenv==0.21.0
